chore: Add support for NEXT_PUBLIC_SUBDIR environment variable
This commit adds support for the NEXT_PUBLIC_SUBDIR environment variable in the next.config.js file. The variable is used to specify a subdirectory for the application. By including this feature, we enhance the flexibility of the application, allowing it to be deployed in different subdirectories. This change improves the overall configurability of the application.