refactor: Update next.config.js to use NEXT_PUBLIC_SUBDIR for assetPrefix
This commit updates the next.config.js file to use the NEXT_PUBLIC_SUBDIR environment variable for the assetPrefix. Previously, the assetPrefix was set to the repository name. By using the environment variable, we enhance the flexibility of the application, allowing it to be deployed in different subdirectories. This change improves the overall configurability and maintainability of the application.